refactor: make component for sparql search form
